working
git-svn-id: svn+ssh://beastie.cs.ua.edu/var/www/svn/scam@401 38ebec9f-8de8-411d-bf7c-b5c56b76c706

+++ util.c
@@ -147,7 +147,7 @@ ProcessOptions(int argc, char **argv)
                 break;
             case 'm':
                 {
-                MemorySize *= atoi(arg);
+                MemorySize *= atof(arg);
                 argUsed = 1;
                 }
                 break;
